Renaming the the page files
